refactor(analysis): carve the overlay family out of the analysis_sql façade
Clip, intersect and select by location move to analysis_sql/overlay.py with
MASK_SUBDIVIDE_MAX_VERTICES and the INTERSECT output-column names. These three
belong together because they genuinely share machinery: the drawn-mask and
mask-layer handling, the ST_Dimension = 2 polygonal guard, and the _mask_pieces
subdivide path that both the clip join and the intersect pairs query build on.

render_geometry_expr stays in the façade as the cross-family dispatcher, but
its clip and select_by_location branches become render_clip_expr and
render_select_by_location_expr in overlay, and the dispatcher calls them. That
is what keeps render_select_by_location_count — which must render the SAME
predicate the preview does, or the count under the map describes a different
set from the features on it — reaching that predicate through its own family
rather than back through the façade, which would be an import cycle.

A spatial join also reads a second layer and is NOT here. It adds columns and
hands the source geometry back untouched, so it shares none of the three things
above; it gets its own module later in this branch.

Byte-identity: all 94 cases pass. It caught the one mistake in this commit — a
dropped ST_MakeValid in the clip WHERE while the branch was being lifted into a
function, which produces valid SQL, a plausible-looking predicate, and a
different row set on any dataset holding a self-intersecting ring.

The façade drops to 677 lines, so analysis_sql leaves _MODULE_LOC_CAPS: nothing
in the package reaches the 1000-line inclusion threshold. The entry's narrative
is rewritten rather than deleted — it was the record of why the module must
never be split by caller, which outlives the ratchet.
